Updated pause popup dialog (made to child window). Other minor changes.
Issues:
- On "Return To App": sometimes opened app will not maximize correctly.
- If you press guide button while aut script is running, sometimes the
openedApp is visible but it is in pause state.
-- Add SCRIPT_RUNNING state to indicate when the aut script is running.
Could wait for aut script to terminate

+++ FilePlayer_Desktop/ViewModels/ItemListViewModel.cs
@@ -114,11 +114,13 @@ namespace FilePlayer.ViewModels
                             SetControllerState("ITEM_PLAY");
                             break;
                         case "CLOSE_APP":
-                            appProc.Kill();
+                            if(!appProc.HasExited)
+                                appProc.Kill();
                             SetControllerState("ITEMLIST_BROWSE");
                             break;
                         case "CLOSE_ALL":
-                            appProc.Kill();
+                            if(!appProc.HasExited)
+                                appProc.Kill();
                             
                             break;
                     }
